Reverted UnnecessaryAbstractClass #534 because of a bug
This refactored rules/style test cases.
This separates positive and negative test cases for rules. (#527)
