+++ detekt-rules/src/test/kotlin/io/gitlab/arturbosch/detekt/rules/KT.kt
@@ -46,9 +46,11 @@ enum class Case(val file: String) {
 	NestedLongMethods("/cases/NestedLongMethods.kt"),
 	FeatureEnvy("/cases/FeatureEnvy.kt"),
 	UnreachableCode("/cases/UnreachableCode.kt"),
-	UnnecessaryAbstractClass("/cases/UnnecessaryAbstractClass.kt"),
+	UnnecessaryAbstractClassPositive("/cases/UnnecessaryAbstractClassPositive.kt"),
+	UnnecessaryAbstractClassNegative("/cases/UnnecessaryAbstractClassNegative.kt"),
 	UtilityClasses("/cases/UtilityClasses.kt"),
-	Serializable("/cases/Serializable.kt"),
+	SerializablePositive("/cases/SerializablePositive.kt"),
+	SerializableNegative("/cases/SerializableNegative.kt"),
 	SuppressedElements("/SuppressedByElementAnnotation.kt"),
 	SuppressedElementsByFile("/SuppressedElementsByFileAnnotation.kt"),
 	SuppressedElementsByClass("/SuppressedElementsByClassAnnotation.kt"),
