+++ detekt-rules/src/test/resources/cases/SerializablePositive.kt
@@ -27,19 +27,3 @@ class IncorrectSerialVersionUID3 : Serializable {
 }
 
 class NoSerialVersionUID : Serializable // violation
-
-class CorrectSerializable1 : Serializable {
-	companion object {
-		const val serialVersionUID = 1L
-	}
-}
-
-class CorrectSerializable2 : Serializable {
-	companion object {
-		const val serialVersionUID: Long = 1
-	}
-}
-
-class NoSerializableClass
-
-interface SerializableInterface : Serializable
