+++ detekt-rules/src/test/resources/cases/UnnecessaryAbstractClassNegative.kt
@@ -1,0 +1,9 @@
+@file:Suppress("unused")
+
+package cases
+
+abstract class AbstractClassOk {
+
+	abstract val i: Int
+	fun f() { }
+}
