+++ detekt-rules/src/test/resources/cases/UnnecessaryAbstractClassPositive.kt
@@ -1,5 +1,3 @@
-@file:Suppress("unused")
-
 package cases
 
 abstract class OnlyAbstractMembersInAbstractClass { // violation: no concrete members
@@ -11,12 +9,6 @@ abstract class OnlyAbstractMembersInAbstractClass { // violation: no concrete me
 abstract class OnlyConcreteMembersInAbstractClass { // violation: no abstract members
 
 	val i: Int = 0
-	fun f() { }
-}
-
-abstract class AbstractClassOk {
-
-	abstract val i: Int
 	fun f() { }
 }
 
@@ -32,8 +24,4 @@ interface AbstractInterface {
 	abstract class NestedAbstractClass { // violation: no abstract members
 		fun f() { }
 	}
-}
-
-abstract class AbstractClassWithPrimaryConstructorConcretePropertyOk(val i: Int) {
-	abstract fun f()
 }
